requirements.txt
cython
plyfile
ujson
pycocotools
cffi
ninja
black
docformatter
setproctitle
meshplex
OpenEXR
vispy>=0.6.4
yacs>=0.1.8
tabulate
pytest-runner
pytest
ipdb
tqdm
numba
mmcv-full
imagecorruptions
pyassimp==4.1.3 
pypng
imgaug>=0.4.0
albumentations
transforms3d
open3d
fvcore
tensorboardX
einops
pytorch3d
timm==0.8.13
glfw
imageio
imageio-ffmpeg
PyOpenGL 
PyOpenGL_accelerate
chardet
thop
loguru
fairscale
deepspeed
pyro-ppl==1.4.0
onnx==1.8.1
onnxruntime==1.8.0
onnx-simplifier==0.3.5
